Fixed a few issues with the application
I have fixed a small issue that would show an exception when showing the App component, due to the call to the llm asking for a question. I have deleted that part since we will not use it in the future and the error disappeared.

More importantly, there was an error that would cause the game to show an exception if the user enters the game before the wikidata information has loaded into the database. Now, while the data is loading, the game will try to get the round and if it fails it will keep showing the loading view to the user.